refactor: mv `get_referenced_exports` and `get_condition` from ModuleDependency to Dependency
