+++ crates/rspack_plugin_javascript/src/dependency/commonjs/require_resolve_dependency.rs
@@ -56,6 +56,14 @@ impl Dependency for RequireResolveDependency {
   fn dependency_debug_name(&self) -> &'static str {
     "RequireResolveDependency"
   }
+
+  fn get_referenced_exports(
+    &self,
+    _module_graph: &ModuleGraph,
+    _runtime: Option<&RuntimeSpec>,
+  ) -> Vec<ExtendedReferencedExport> {
+    vec![]
+  }
 }
 
 impl ModuleDependency for RequireResolveDependency {
@@ -77,14 +85,6 @@ impl ModuleDependency for RequireResolveDependency {
 
   fn set_request(&mut self, request: String) {
     self.request = request;
-  }
-
-  fn get_referenced_exports(
-    &self,
-    _module_graph: &ModuleGraph,
-    _runtime: Option<&RuntimeSpec>,
-  ) -> Vec<ExtendedReferencedExport> {
-    vec![]
   }
 }
