+++ crates/rspack_plugin_javascript/src/dependency/worker/mod.rs
@@ -53,6 +53,14 @@ impl Dependency for WorkerDependency {
   fn dependency_debug_name(&self) -> &'static str {
     "WorkerDependency"
   }
+
+  fn get_referenced_exports(
+    &self,
+    _module_graph: &ModuleGraph,
+    _runtime: Option<&RuntimeSpec>,
+  ) -> Vec<ExtendedReferencedExport> {
+    vec![]
+  }
 }
 
 impl ModuleDependency for WorkerDependency {
@@ -66,14 +74,6 @@ impl ModuleDependency for WorkerDependency {
 
   fn set_request(&mut self, request: String) {
     self.request = request;
-  }
-
-  fn get_referenced_exports(
-    &self,
-    _module_graph: &ModuleGraph,
-    _runtime: Option<&RuntimeSpec>,
-  ) -> Vec<ExtendedReferencedExport> {
-    vec![]
   }
 }
