+++ packages/rspack/tests/cases/parsing/issue-5197/index.js
@@ -1,0 +1,7 @@
+require("./a.mjs");
+
+it("should retain require callee", () => {
+	const fs = require("fs");
+	const content = fs.readFileSync(__filename, "utf-8");
+	expect(content).toContain(`require('./in-exists')`);
+});
